fix(vector): three crash fixes for matmul and comparison ops
Bug 1: extract_element_512 panicked on out-of-bounds word index when
geometry produced byte offsets >= 64. Add bounds checks returning 0
for elements beyond the 512-bit vector.

Bug 2: bf16 config word 0x3d (variant=1, element-wise mode) was
unrecognized. Add bf16 element-wise geometry: 16 channels of 1x2
dot products.

Bug 3: VGE/VLT comparison mask assembly panicked on shift-left overflow
when lanes_per_half=32 (8-bit elements). Widen to u64 before shifting.

Generated using Claude Code.

+++ src/interpreter/execute/vector.rs
@@ -3469,8 +3469,10 @@ impl VectorAlu {
                     let mask_lo = Self::condense_comparison_mask(&lo, op.element_type);
                     let mask_hi = Self::condense_comparison_mask(&hi, op.element_type);
                     let lanes_per_half = 256 / et.bits() as u32;
-                    let full_mask = mask_lo | (mask_hi << lanes_per_half);
-                    Self::write_scalar_dest(op, ctx, full_mask);
+                    // Use u64 for the combined mask since lanes_per_half can
+                    // be 32 (8-bit elements), which would overflow a u32 shift.
+                    let full_mask = (mask_lo as u64) | ((mask_hi as u64) << lanes_per_half);
+                    Self::write_scalar_dest(op, ctx, full_mask as u32);
                 } else {
                     Self::write_wide_vec_dest(op, ctx, result);
                 }
@@ -3486,8 +3488,10 @@ impl VectorAlu {
                     let mask_lo = Self::condense_comparison_mask(&lo, op.element_type);
                     let mask_hi = Self::condense_comparison_mask(&hi, op.element_type);
                     let lanes_per_half = 256 / et.bits() as u32;
-                    let full_mask = mask_lo | (mask_hi << lanes_per_half);
-                    Self::write_scalar_dest(op, ctx, full_mask);
+                    // Use u64 for the combined mask since lanes_per_half can
+                    // be 32 (8-bit elements), which would overflow a u32 shift.
+                    let full_mask = (mask_lo as u64) | ((mask_hi as u64) << lanes_per_half);
+                    Self::write_scalar_dest(op, ctx, full_mask as u32);
                 } else {
                     Self::write_wide_vec_dest(op, ctx, result);
                 }
